refactor: 💡 rename RenderForSizes3 to RenderForSizes
