--- src/Molecules/FlexTable/shared/RenderForSizes3/index.ts
@@ -1,3 +1,0 @@
-import { RenderForSizes, getScreenMedia } from './RenderForSizes';
-
-export { RenderForSizes, getScreenMedia };
