+++ src/Molecules/FlexTable/shared/index.ts
@@ -1,7 +1,6 @@
 import { CellInlineContainer } from './CellInlineContainer';
 import { ColumnProvider } from './ColumnProvider';
 import { RenderForSizes, getScreenMedia } from './RenderForSizes';
-import { RenderForSizes as RenderForSizes3 } from './RenderForSizes3';
 import * as constants from './constants';
 import { StyledTruncateTooltip } from './StyledTruncateTooltip';
 
@@ -9,7 +8,6 @@ export {
   constants,
   ColumnProvider,
   RenderForSizes,
-  RenderForSizes3,
   getScreenMedia,
   CellInlineContainer,
   StyledTruncateTooltip,
